requirements.txt
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
scikit-learn==1.5.1
matplotlib==3.9.0
seaborn==0.13.2
jupyter==1.0.0
pyyaml==6.0.2
